backend/requirements.txt
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==3.2.2
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.3.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
hyperframe==6.1.0
idna==3.11
kombu==5.6.1
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
passlib==1.7.4
portalocker==2.10.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
qdrant-client==1.6.4
redis==5.0.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wrapt==2.0.1
